move to util package
