+++ logpusher/src/main/java/org/weiboad/ragnar/util/Toolbox.java
@@ -1,4 +1,4 @@
-package org.weiboad.ragnar.common;
+package org.weiboad.ragnar.util;
 
 import org.apache.commons.cli.*;
 import org.jetbrains.annotations.NotNull;
